requirements.txt
celery~=5.3.4
Django~=5.0.2
django_components==0.81
django-environ~=0.11.2
django-extensions~=3.2.3
django-filter~=23.5
django-htmx~=1.17
django-webpush~=0.3.6
django-widget-tweaks~=1.5.0
gunicorn~=21.2.0
psycopg2~=2.9.8
pytailwindcss~=0.2.0
whitenoise~=6.5.0
inotify~=0.2.10
redis~=5.0.1
